demo_backend_dev/requirements.txt
Flask==2.0.2
Turbo-Flask==0.8.0
gunicorn==20.1.0
invoke==1.7.1
requests==2.28.1

google-auth==2.3.2
